Adjusting BlockInfoHandler
Added new method to check if a player can interact (f.ex. break) width a
block. Ment to collect all the logic in listeners for this to one
method.
isProtected() is now using getOwnerId(Block) to check if a block is
protected.
Added new method getOwnerID(int, int, int, String) to allow checking
with coords and world name.